// frontend/src/shared/api/generated/models/ReportPayload.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ReportShareContext } from './ReportShareContext';
export type ReportPayload = {
    sessionId: number;
    raw?: (Record<string, any> | null);
    percentiles?: (Record<string, any> | null);
    style?: (Record<string, any> | null);
    lfi?: (Record<string, any> | null);
    visualization?: (Record<string, any> | null);
    sessionDesigns?: null;
    analytics?: (Record<string, any> | null);
    learningSpace?: (Record<string, any> | null);
    enhancedAnalytics?: (Record<string, any> | null);
    notes?: (Record<string, any> | null);
    owner?: (Record<string, any> | null);
    shareContext?: (ReportShareContext | null);
    responsibleUseNotice?: (string | null);
};